Unlpimage/requirements.txt
PySimpleGUI==4.60.4
Pillow==9.3.0
pandas==1.5.2
matplotlib==3.6.2
notebook==6.5.2
wordcloud==1.9.2
